+++ example/seat/seatInfoSync.js
@@ -1,17 +1,17 @@
 /**
  * seatInfoSync
  */
-const client = require('../../lib');
+const { Surfrock } = require('../../lib');
 
 async function main() {
-    const authClient = new client.auth.ClientCredentials({
+    const authClient = new Surfrock.auth.ClientCredentials({
         domain: process.env.API_AUTHORIZE_SERVER_DOMAIN,
         clientId: process.env.API_CLIENT_ID,
         clientSecret: process.env.API_CLIENT_SECRET,
         scopes: [],
         state: ''
     });
-    const seatService = new client.service.seat.SeatService(
+    const seatService = new Surfrock.service.seat.SeatService(
         {
             endpoint: process.env.API_ENDPOINT,
             auth: authClient
@@ -23,8 +23,8 @@ async function main() {
     );
     const result = await seatService.seatInfoSync({
         kgygishCd: 'SSK000',
-        yykDvcTyp: client.factory.service.seat.seatInfoSync.ReserveDeviceType.EntertainerSitePC,
-        trkshFlg: client.factory.service.seat.seatInfoSync.DeleteFlag.False,
+        yykDvcTyp: Surfrock.factory.service.seat.seatInfoSync.ReserveDeviceType.EntertainerSitePC,
+        trkshFlg: Surfrock.factory.service.seat.seatInfoSync.DeleteFlag.False,
         kgygishSstmZskyykNo: '118124',
         kgygishUsrZskyykNo: '124',
         jeiDt: '2017/03/02 10:00:00',