+++ src/index.ts
@@ -1,4 +1,1 @@
-export { SurfrockFactory as factory } from '@surfrock/factory';
-export * as service from './service/index';
-export * as auth from './auth/index';
-export { RequestError } from './transporters';
+export * as Surfrock from './surfrock';