+++ src/surfrock.ts
@@ -1,0 +1,4 @@
+export { SurfrockFactory as factory } from '@surfrock/factory';
+export * as service from './service/index';
+export * as auth from './auth/index';
+export { RequestError } from './transporters';